Change implementation of `containerStackView`.
In order to reset the `containerStackView`, we were relying on a store property and a computed one.
This way handy to get the functionality of a lazy property that can be reinitialized.
We now improved this design by manipulating the arranged subviews directly. This works since we can simply reuse the same instance.